// src/app/models/opportunity.ts
export class Opportunity extends BaseModel
{
    id: string;
    accountId: string;
    accountName: string;
    amount: number;
    amountUsdollar: number;
    campaignId: string;
    contactId: string;
    contactName: string;
    createdBy: string;
    currencyId: string;
    dateClosed: string;
    dateCreated: Date;
    dateModified: Date;
    deleted: boolean;
    description: string;
    leadSource: string;
    modifiedUserId: string;
    name: string;
    nextStep: string;
    opportunityType: string;
    probability: string;
    salesRep: string;
    salesRepId: string;
    salesStage: string;
    opportunityNo: string;
    
    setData(data) {
        this.setString(data, 'id');
        this.setString(data, 'accountId');
        this.setString(data, 'accountName');
        this.setFloat(data, 'amount');
        this.setFloat(data, 'amountUsdollar');
        this.setString(data, 'campaignId');
        this.setString(data, 'contactId');
        this.setString(data, 'contactName');
        this.setString(data, 'createdBy');
        this.setString(data, 'currencyId');
        this.setLocalDate(data, 'dateClosed');
        this.setDate(data, 'dateCreated');
        this.setDate(data, 'dateModified');
        this.setBoolean(data, 'deleted');
        this.setString(data, 'description');
        this.setString(data, 'leadSource');
        this.setString(data, 'modifiedUserId');
        this.setString(data, 'name');
        this.setString(data, 'nextStep');
        this.setString(data, 'opportunityType');
        this.setFloat(data, 'probability');
        this.setString(data, 'salesRep');
        this.setString(data, 'salesRepId');
        this.setString(data, 'salesStage');
        this.setString(data, 'opportunityNo');
    }
}